feat: support primitive arrays

+++ BinaryWizard/BinaryWizard.Sample/SampleEntity.cs
@@ -20,7 +20,7 @@ namespace BinaryWizard.Sample;
 public partial class SampleEntity {
     public SampleVector Vector;
 
-    [BinaryArray(Size = 2)] public byte[] Ids;
+    [BinaryArray(Size = 2)] public int[] Ids;
 }
 
 [BinarySerializable]
